+++ src/plugins/intel_gna/tests/functional/shared_tests_instances/behavior/ov_executable_network/compiled_model_base.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "behavior/ov_executable_network/exec_network_base.hpp"
+#include "behavior/ov_executable_network/compiled_model_base.hpp"
 
 using namespace ov::test::behavior;
 namespace {
